+++ src/PeakPresence/ConfigHandler.cs
@@ -12,6 +12,10 @@ public static class ConfigHandler
 
     public static ConfigEntry<bool> UseDetailedDetails { get; private set; } = null!;
 
+    public static ConfigEntry<bool> ShowAliveStatus { get; private set; } = null!;
+    public static ConfigEntry<bool> ShowHeight { get; private set; } = null!;
+    public static ConfigEntry<bool> AbbreviateHeight { get; private set; } = null!;
+
     public static ConfigEntry<string> LogoImageKey { get; private set; } = null!;
     public static ConfigEntry<string> ShoreImageKey { get; private set; } = null!;
     public static ConfigEntry<string> TropicsImageKey { get; private set; } = null!;
@@ -24,6 +28,8 @@ public static class ConfigHandler
     public static void Initialize(ConfigFile configFile)
     {
         config = configFile;
+
+
         DiscordAppID = config.Bind("General", "DiscordAppID", "1408478763682894045", "The Discord Application ID to use for Rich Presence.");
 
         ForcedLanguage = config.Bind("General", "ForcedLanguage", "", new ConfigDescription(
@@ -32,6 +38,10 @@ public static class ConfigHandler
         ));
 
         UseDetailedDetails = config.Bind("General", "UseDetailedDetails", false, "Whether to use detailed presence information instead of \"In Game: {location}\"");
+
+        ShowAliveStatus = config.Bind("General", "ShowAliveStatus", true, "Whether to show if the player is alive or dead in the presence.");
+        ShowHeight = config.Bind("General", "ShowHeight", true, "Whether to show the player's height in the presence.");
+        AbbreviateHeight = config.Bind("General", "AbbreviateHeight", true, "Whether to abbreviate the height measurement (e.g. 1500 m -> 1.5 km).");
 
         LogoImageKey = config.Bind("Images", "LogoImageKey", "logo", "The key for the logo image.");
         ShoreImageKey = config.Bind("Images", "ShoreImageKey", "shore", "The key for the shore image.");